Fix errors and warning when compiling with Visual C++ 2010 64 bit.
* One warning remains. A type name in BayesianNetwork.cpp is too long. The
  compiler shortens it. This seems to have no further negative effect.

* I did not test the library on Windows though. This patch just removes
  compiler errors and warnings.

+++ src-lib/DirichletProcessNode.cpp
@@ -64,7 +64,7 @@ namespace cpprob
     for (auto count = counters.begin(); count != counters.end(); ++count)
     {
       // The prior
-      p = count->second;
+      p = static_cast<float>(count->second);
 
       // The factors of the posterior update
       value_ = count->first;
